cargo: Fix tracking with aliases and no refs
It was reported that `bst track` raised an error when the url had
aliases and no refs were specified https://gitlab.com/BuildStream/bst-external/-/issues/53

This appears to have been caused by not marking the url in `configure`
when there are no refs.

This commit fixes it by always marking the url in `configure`.