docker: fix authentication when using .netrc
From the requests documentation

> The netrc file overrides raw HTTP authentication headers set with headers=.

Which means that when the .netrc file contains the credentials for a docker
registry, it will override the bearer token authentication that the plugin
is setting using a raw "Authorization" header.

Instead, we create a new BearerAuth class and pass it to requests to use
for authenticating to the registry.