cargo: Add support for git dependencies
Cargo supports specifying sources as originating from git repositories
instead of registries like crates.io.

This adds support for staging and fetching the dependencies. It adds and
explicitly "kind" key to the crate entries in elements that default to
"registry" for backward compatibility.

The staging emulates cargo vendoring by adding a config entry for each
git repository and generating a checksum file for the crate.

Closes #10