Adding simple_mirror source mirror plugin
Since we've added the SourceMirror feature in BuildStream 2.2, it makes
sense to add this mirror plugin upstream.

This is especially helpful as we can rely on loading stable
buildstream-plugins via the `pip` plugin origin for loading source
mirror plugins, avoiding a cyclic dependency issue if you want to
use mirror plugins to load plugins via the `junction` plugin origin.