tox.ini: Conditionally wrap the bzr command if it's present on the host
Since bzr is not implemented with breezy, and breezy uses a python library,
we cannot execute bzr on some platforms without installing the breezy python
libs into the tox venv, and we cannot install breezy into the venv easily
since it is installing from source, and using things like cython and rust.

Instead, we simply wrap host bzr with a script which overrides the PATH
variable, so that python ignores the venv when entering the bzr code, this
allows us to keep our tox environment clean for buildstream and tests, while
testing the host installation of bzr.