sources/docker.py: Use ReadableTarInfo when creating member file sets
The list of TarInfo received from tar.getmembers() is returned from
_get_extract_and_remove_files() to the caller and used to specify the
members to be extracted from the tar file. If the "tarinfo" parameter
is not set to ReadableTarInfo when the member list is created, its
file mode modifications would not apply to those instances of TarInfo.

As a result, the logic in ReadableTarInfo did not actually apply to
the file extraction process and file modes remained the same. This
results in errors such as the following when staged files cannot be
read by the running user:

  [00:10:04] FAILURE [0e4ec5fd] images/redhat-ubi8-x86_64.bst: Failed to capture tree /home/user/.cache/buildstream/tmp/staging-temppou9rkp9: code: 13
  message: "System error in `merklize()` for path \"/home/user/.cache/buildstream/tmp/staging-temppou9rkp9\": std::system_error exception thrown at [buildboxcommon_merklize.cpp:1125] [system:13], errMsg = \"Failed to open path \"\"etc/gshadow\"\"\", errno : Permission denied"

Setting tarinfo=ReadableTarInfo in _get_extract_and_remove_files()
solves this issue.