sources/docker.py: Process whiteout files before extracting
The previous approach relied on removing files which were already
extracted in a previous step, which was susceptible to permission
and ownership issues.

For example, the following error is seen when the the permissions
of the parent doesn't allow removing a file:

  [00:00:06] FAILURE [0e4ec5fd] images/redhat-ubi8-x86_64.bst: docker source at images/redhat-ubi8-x86_64.bst [line 4 column 2]: Error staging source: [Errno 2] No such file or directory: '/home/coder/.cache/buildstream/tmp/staging-tempi542ewmq/root/buildinfo/'

Processing layer members before extracting them should avoid such
issues by never having to remove any files. It should also result
in significantly reduced disk IO when many whiteout files are
present.

The downside is slightly higher memory usage to store the member
list of all layers before extracting them, but the difference
should not be significant.